+++ torchtune/models/llama2/_model_builders.py
@@ -12,7 +12,7 @@ from torchtune.modules import TransformerDecoder
 from torchtune.models.llama2._tokenizer import Llama2Tokenizer
 from torchtune.modules.peft import LORA_ATTN_MODULES
 from torchtune.data._prompt_templates import _TemplateType
-from torchtune.data._utils import _get_prompt_template
+from torchtune.data._prompt_templates import _get_prompt_template
 
 
 """
