+++ torchtune/models/phi3/_model_builders.py
@@ -8,7 +8,7 @@ from torchtune.modules.peft import LORA_ATTN_MODULES
 from functools import partial
 from torchtune.modules.tokenizers import parse_hf_tokenizer_json
 from torchtune.data._prompt_templates import _TemplateType
-from torchtune.data._utils import _get_prompt_template
+from torchtune.data._prompt_templates import _get_prompt_template
 
 
 """
